Ensure component tests don't run as root user on Linux
On linux machines, we don't have a docker desktop, as in Mac or Win, so
the cache files are being created as the default user on the service
container which is root.

This causes the `.go` folder to be owned my root which then forces us to
either change the owner directory, remove the `.go` or run the tests as
root.

This change ensures that the userID and user groupID running the component tests on the
host machine are passed into the service container, ensuring that the
`.go` is owned by the same user and not root.