5758 - Create an HTTPClient with the DEFAULT_REQUEST_TIMEOUT
The current value of the httpClient is set to 10s.
Creating an HTTPClient with a default timeout will allow us to avoid a
GraphQL connection timeout when requesting data for very large datasets.

The reason this should prevent the timeout is because the GraphQL POST
query is done via the normal httpClient.

The error we are seeing in Kibana:

./dp-observation-api
```
  write tcp 172.17.0.17:50838->10.30.137.41:8182: write: broken pipe
```

dp-hierarchy-builder
```
  message": "write tcp 172.17.0.16:36782->10.30.136.54:8182: write: connection timed out
     event: "error from graph DB"
```

dp-cantabular-csv-export
```
   event: "failed to handle message"
  errors: "failed to upload .csv file to S3 bucket: failed to stream csv data:
           transform error: error decoding end of json object: error decoding json token:
           context deadline exceeded (Client.Timeout or context cancellation while reading body"
 ```

Resolves: 5758